Do some more stuff, you filthy heretics :P
I'm not really happy with the direction the devs are taking this. And in
particular, they should *please* consider removing their weird debug
stuff when doing a release. There's some "loop0sub0" weirdness displayed
next to the clock now, monsters sometimes display weird "combo@10" or
whatever strings now, and to top it off, if a nurse is damaged the game
now displays "I smell nurse c", an obvious sign that this is a beta
version! This kind of sloppiness is not what I'm used to seeing in new
releases of this game. Dear devs, you can do better than that, please
put some effort into it and try to polish your code before making a
release.