Can recharge rod of domination
That was a really stupid design decision. You can reserve the spellbook at the magic spell writer *AND* recharge the hell out of it, so why shouldn't you be able to do the same with the rod that is much more difficult to get to begin with (requiring you to either get very lucky to find one, because shops never seem to stock it, or farming 20 medals for every 4 charges you want)? So there.